requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
python-dotenv
pymysql
sqlalchemy
mysql-connector-python
xgboost==3.2.0
-e .